refactor(checker): split object literal diagnostic anchors
